Added insert program
This time I added an insertion program. This is not yet finished as it needs to update the JSON database for it to be perfectly finished. But it takes the items, finds the best slots to put the items in the storage and warns if some items couldn't find space.