Modified search and registries to handle mods.
The registries in the config file are now in a list of paths pointing to the json files containing the IDs of the blocks. There can be multiple mods per file, one mod per file, as long as everything is mentioned in that file, the blocks can be found by the network.