Refined search function to use NBT Data
The search function now correctly differentiates the items using the same id with different internal data. The database structure has been changed to accomodate this fact. In an item section, there is now a potential array in a nbt section of NBT found for this id, and the information about the items stacks in a stacks section.