Added extract program
Added extraction program which allows, with the use of prompted user input (and autocompletion for ids), to extract items from the storage to the netowrk output inventory mentioned in the config gile. It also updates the JSON database so the inventory program doesn't have the be run again. Also, fixed some bugs with the way the insert program interacted with the JSON database and moved JSON database edit function to the utils.lua file.