Extraction prompt for choosing variant
Added a new interactive prompt that will interrupt the extraction process to ask the user to choose a variant of the item that they chose, should it exist. (For example, potions). In the case where both unique and default items exist on the storage (for example, named items and their default counterpart), all unique items will have their rows, and a row with the NBT value DEFAULT will also be present to give the user the ability to choose the default items. Also, fixed a lot of bugs and tried to temporarily improve insertion speed by shuffling the storage names list so that it doesn't take the same way each time, which would slow down after each insertion. Still, marking the storages as full is very much necessary.