Performance improvements
Changed a huge part of the codebase to add optimizations especially for the insert program which had a huge performance flaw. Also decided that whenever the inventory program is ran, the empty spaces in the inventories should also be cached so that the search for empty stacks when inserting is much faster.