Added timing of programs
Added two functions to the utils files used to get how much time has passes between two function. The time a program took to execute is then display in a TIMER log to the user at the end of the program, if ended successfully. Also, quick bugfix on the search function for items that aren't in storage.