Optimized extraction even more
Removed a useless part of the extraction code and considered only two types of cases : If a stack has not enough OR just enough items to satisfy the call to the function OR the stack has too many items. In the first case, the full stacks are always considered 'just enough' and then the rest falls in one of the two cases. Goes faster now :)