Fixed cursor issues
Fixed some issues with the cursor when having to select from the simple list. If the list had 20 rows and the list had 20 entries, the cursor would get put on position 0 instead of 1 since the calculation was made with modulo. This is now fixed. It's a bit ugly but hey, it's a terminal.