Added details program
Added a details program that is only accessible through a simple search. When the user is browsing the search results, they can choose to either quit or get details about the item that the cursor points to. Then, they are shown the details of the item and here they can either start the extraction process of said item or go back to the base search list.